Remove existing Action enum support (first pass).
