+++ Sources/TCAComposerMacros/Extensions.swift
@@ -123,3 +123,13 @@ extension KeyPathExprSyntax {
     self.init(components: components)
   }
 }
+
+extension DeclModifierSyntax {
+  var isNeededAccessLevelModifier: Bool {
+    switch self.name.tokenKind {
+    case .keyword(.public): return true
+    case .keyword(.package): return true
+    default: return false
+    }
+  }
+}
